requirements.txt
pdfplumber>=0.9
PyMuPDF>=1.20  # PyMuPDF (imported as 'fitz')
python-docx>=0.8
python-pptx>=0.6
Pillow>=9.0
